[LibOS,PAL,Docs] Remove `loader.argv0_override` manifest option
This manifest option was always set to `libos.entrypoint` anyway (or
abused in one of the LibOS ABI stack tests). Instead, when there is no
argv-handling manifest option specified, Gramine sets
`argv = [ libos.entrypoint ]` (or `argv = [ loader.entrypoint ]` if
there is no LibOS, as happens in PAL tests). Note that this only happens
in the first process; child processes get argv from the parent as-is.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>