[LibOS/regression] Fix `test_212_shebang_test_script` to run with bash
Bash first resolves the path to be absolute and then passes it as
pathname (contrary to dash, which does not do it).

Dash is a default `/bin/sh` on Ubuntu, so the shebang test previously
used a relative pathname. But on CentOS/RHEL/Fedora, Bash is the default
`/bin/sh`, and thus the shebang test failed. With this commit, we
introduce a Regex that works both for relative and absolute pathnames.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>