[PAL/Linux-SGX] Bump the max number of enclave threads to 4096
Modern Intel CPUs have a large number of cores, and modern workloads
(such as PyTorch) allocate a large number of threads (e.g., 4 threads
per each CPU core). This exceeds the previous limit on the number of
threads (1024) and leads to boostrap errors like this:

    error: There are no available TCS pages left for a new thread!
    Please try to increase sgx.thread_num in the manifest.

This commit bumps the limit to 4096 and fixes several potential
stack-overflow issues along the way.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>