[tests] Use #include mbedtls headers in the attestation test
Previously, we inlined the mbedtls relevant definitions due to no access
to mbedtls headers (mbedtls is built as a Meson suproject while the
attestation test file was compiled using Make at that time).

As this test is built using Meson now, we should use #include mbedtls
headers instead of the definitions.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>