[common] Init new protected file when file size is zero
Currently, Gramine does not close a file opened inside an encrypted
mount but not explicitly closed by the application. No data is ever
written to the host in this case. A subsequent opening of this file
(e.g., on the application restart) will be broken since it's considered
as operating on an existing file however the file size is actually zero
and the encrypted file header is also missing.

This patch initializes a new protected file instead of an existing one
when the file size is found out to be zero.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>