[LibOS,PAL,Docs] Introduce etc passthrough
The first `etc` passthrough file is the `/etc/hostname`. Gramine reads
the option to global PAL state, and LibOS uses it to create a pseudo file
with its content. Gramine sanitizes hostname. It requires that it is a valid
domain. This is a difference from Linux, as Linux accepts any hostname value.
That said, Linux doesn't assume that the root user tries to exploit it through
hostname, and Gramine should.

Gramine uses pseudofs for `etc` passthrough.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>