[LibOS] Fix an UAF on PAL event in `clone_implementation_wrapper`
`clone_implementation_wrapper` uses a PAL event to notify parent thread
that it's done initializing. After the parent thread sees the event set,
it frees all temporary resources including that event. This could lead
to a subtle race, when the child thread marks the even as set, but is
not yet done using it (because e.g. internal implementation of setting
the event needs to wake up potential waiters, so it needs to read some
fields from the event struct).
Usually PAL events are wrapped in some objects and their lifetime is
tied to lifetime of that object - if we have a reference to the object,
we can safely use the event. It was not the case in
`clone_implementation_wrapper`, where the lifetime of the event was
incorrectly tied to whether the event itself is set or not. This commit
solves thiat by adding an atomic variable denoting whether the child
thread is done using the event.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>