[PAL] Remove `PAL_OPTION_CLOEXEC` flag
This flag was used to control `CLOEXEC` flag of host OS fd, and as such,
was Linux specific. It was also not really needed - Linux and Linux SGX
PALs always set it on host fd anyway.
This commit also adds all missing `CLOEXEC` flag usages in Linux and
Linux SGX PALs.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>