squash! [LibOS,PAL,Docs] Introduce etc passthrough
!TODO use this commit message:

[LibOS,PAL,Docs] Introduce `etc` emulation (currently only 'hostname')

Gramine obtains information from a host, sanitizes it, and stores it in
the global PAL state. Later, LibOS uses it to create a pseudo file
(using pseudofs filesystem).

The difference between Linux and Gramine is that Gramine expected
the hostname to be a valid domain - as we assume the host is untrusted.
In the case of Linux, it doesn't presume that the root user
tries to exploit it through a malicious hostname.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>