[PAL/Linux-SGX] Remove insecure `enclave_untrusted.c`
This file contained implementation of `malloc_untrusted`, which allowed
in enclave code to allocate untrusted memory of arbitrary size in small
granularity (contrary to page size in `ocall_mmap_untrusted`).
The problem with its implementation was it used `slabmgr.h` which holds
all metadata inline - in this case in untrusted memory. This was
trivially exploitable by malicious host OS.

`malloc_untrusted` was used in one place only and this commit adds a
simple untrusted memory allocator for this specific kind of objects.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>