squash! [LibOS,PAL] Extend extra runtime configuration to include hostname
!!TODO use this commit message:

[LibOS,PAL] Extend extra runtime configuration to include hostname

Gramine exposes hostname information to the app via `nodename` field returned
by the `uname` syscall. Note that currently Gramine does *not* create an
`/etc/hostname` file as no apps seem to read this file directly.

To expose the hostname, Gramine obtains it from the host, sanitizes it, and
stores it in the global PAL state.

The difference between Linux and Gramine is that Gramine expects
the hostname to be a valid domain - as we assume the host is untrusted.
(In the case of Linux, it doesn't presume that the root user
tries to exploit it through a malicious hostname.)

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>