[CI-Examples] Add `--test-malicious-quote` arg for ra-tls-mbedtls server
Previously, the server in the `ra-tls-mbedtls` example corrupted the SGX
quote (for testing purposes) when it was started with *any* arguments.
With this commit, the server recognizes only `--test-malicious-quote`
arg and complains on other arguments. This change prevents users from
accidentally corrupting the SGX quote when they mistype something like
`gramine-sgx server dcap`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>