[CI-Examples] Harden Python example
This commit removes the manifest option `sgx.allowed_files` with its
unneeded `/etc/` files. Also the commit cleans up the paths in the
manifest template.

The helper file with "localhost" mappings of IP addresses to local names
is added, required by getaddrinfo() network address translation. It is
mounted as an `/etc/hosts` file and marked as `sgx.trusted_files`.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>