[LibOS] Change real UID/GID if current process is privileged
The `setuid` and `setgid` syscalls set effective UID (`euid`) and
effective GID (`egid`) respectively. But they also set real UID (`uid`)
and real GID (`gid`) if the current process is privileged. This commit
adds this last corner case, which was missing previously.

Since Gramine doesn't have the concept of capabilities, we use the
`uid == 0` check to decide whether the process is privileged.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>