[LibOS] Create file on host corresponding to named UNIX socket
Named UNIX domain sockets are expected to be backed by files.
Previously, Gramine didn't have such backing files. However, some
workloads require this, e.g., two subprocesses spawned in the same
Gramine instance and communicating over a named UNIX socket. A new LibOS
regression test is added that tests this.

This is a poor-man's implementation: LibOS creates a dummy regular file
on the host, so that both subprocesses can see it. This poor
implementatoin is chosen because there is no file synchronization
mechanism available in Gramine.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>