[python] Move from `toml` library to `tomli` + `tomli-w`
`toml` Python library (for parsing and generating TOML files) is buggy
and doesn't support the full spec of TOML 1.0.0. This commit replaces it
with more robust `tomli` (for parsing) and `tomli-w` (for generating).

New LibOS regression test is added that combines all tricky/legacy TOML
syntax, including the checks for TOML-syntax bugs.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>