[PAL/Linux-SGX] Rework SGX driver path resolving
Cleaning up and updating everything to the current state.

Additionally, moved resolving of DCAP/upstream driver paths to the
runtime, so that we can have a single build supporting both the
upstreamed driver and the legacy DCAP one (the OOT DCAP driver and the
upstream driver have the exact same userspace interface, so we can
unify them).

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>