[LibOS,PAL] Support `(E)POLLRDHUP` and treat error events separately
Previously, `(E)POLLRDHUP` was not reported and `(E)POLLHUP` was always
reported together with `(E)POLLERR` due to the limitation of PAL API,
which did not distinguish these conditions. This may cause some
workloads to ignore the fact that the socket was closed by the peer and
hence keep it in the `CLOSE-WAIT` state.

This commit adds support for `(E)POLLRDHUP` and treats the error events
(`(E)POLLERR`, `(E)POLLHUP`, `(E)POLLNVAL`, `(E)POLLRDHUP`) separately.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>