squash! [LibOS] Do not overwrite previous memory in `mmap`
TODO: new commit message below

[LibOS] Do not overwrite previous memory in `mmap`

When `mmap` with `MAP_FIXED` is called on existing (already mapped)
memory, we now first free it and only then allocate new memory at the
given address.
This commit also removes an assumption that `PalVirtualMemoryFree` is
able to free a range of memory with non-mapped holes.

Signed-off-by: Borys Popławski <borysp@invisiblethingslab.com>