[*] debian: Remove python3-protobuf dependency
python3-protobuf changed API at 3.12 (between Debian 10 and 11, and
between Ubuntu 20.04 and 22.04). It made it impossible to ship a _pb2.py
file that would work across this version, which in turn made us to ship
different packages for different distros. But since we don't package
-oot version anymore, we don't need to touch aesmd from python (it was
needed to get EINITTOKEN), we can simply skip this module and remove the
dependency.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>