[LibOS,PAL] Move assigning name to a pipe/UDS from PAL to LibOS
This commit moves the current restriction of pipes + UDSes (that they can connect
to processes only within the same Gramine instance) from PAL layer to LibOS layer.

As a side effect, this commit renames `instance_id` to a more appropriate
`namespace_id` ("Gramine instance" was a term we used to describe a set of
Gramine processes that can communicate with each other, but the better term
for this is "namespace").

Signed-off-by: Ying Liu <ying2.liu@intel.com>