[PAL/Linux-SGX] Try old mmap flags if first mmap on SGX driver failed
Commit "[PAL/Linux-SGX] Add initial EDMM support (only regular pages)"
simplified the initial mmap request to the SGX driver (to allocate EPC
memory). However, this simplification broke the OOT DCAP driver v1.33.2,
found on MS Azure VMs with Ubuntu 18.04: it fails with EACCES.

It turns out the OOT DCAP driver and the upstream driver are not 100%
compatible because of this mmap. Hopefully we won't need to support the
OOT DCAP driver for long, so this commit introduces a hack to revert to
the old mmap flags on failure.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>