[CI] Run AlmaLinux 9 on hosts that support `clone3`
A new version of glibc uses `clone3` to create a thread.
This thread is used to resolve a hostname. AlmaLinux 9 uses
a new version of this library. CI agents need to support
`clone3 syscall` to test Gramine on this distribution,
because Docker uses the kernel from the host.

The `clone3` syscall was introduced in the Linux kernel 5.3.
Besides the kernel, userland utilities also have to support it,
like Docker when creating valid seccomp rules.

To indicate full support of this syscall new `clone3` label has
been introduced.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>