[CI] Run almalinux9 on hosts that support `clone3`
A new version of glibc uses `clone3` to create a thread, to
resolve a hostname. almalinux9 uses a new version of this
library. CI agents need to support `clone3 syscall` to test
Gramine on this distribution.

The `clone3` syscall was introduced in the Linux kernel 5.3.
Besides the kernel also userland utilities have to support it,
like docker to create valid seccomp rules.

To indicate full support of this syscall new `clone3` label has
been introduced.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>