[LibOS] Ignore ancillary data on `recvmsg()`
Previously, Gramine returned `-ENOSYS` on `recvmsg()`/`sendmsg()` and
similar syscalls if it detected the use of `struct msghdr` ancillary
data (`msg_control` field). However, some applications may initialize
this ancillary data for `recvmsg()`, even if it won't be used (e.g.,
ancillary data on TCP/IP may be specified but is ignored) -- this is the
case of `swtpm` software, for example. This commit relaxes Gramine
behavior in such cases on TCP/IP sockets, choosing to ignore ancillary
data instead of failing.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>