[common] Fix POSIX locking on encrypted file in child process
There was a bug in the corner case of POSIX locking: when the child
process acquires a lock on an encrypted file that was not processed yet
by the master process. In this case, the child sends an IPC request to
the master process (since all POSIX locks' operations are centrally
processed in the master process). The master process must lookup the
file first, and only if the lookup is successful, grant the POSIX-lock
request to the child.

The bug was: because encrypted files were *not* updated on the hard disk
upon creation (the file was created but its metadata wasn't flushed),
this confused the master process, and POSIX lock operations failed. This
commit fixes this bug by immediately flushing the metadata of each newly
created encrypted file.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>