[PAL/Linux-SGX] Verify OCALL-obtained time against last RDTSC time
Previously, Gramine did not verify the read-from-Linux-host timestamp
against the last RDTSC time. So it is possible periodically that the
previous timestamp (calculated via RDTSC) is greater than the current
timestamp (read from Linux host via OCALLs), which can lead to the
intermittent failure of LTP test `gettimeofday02` on some
configurations with indication that "time is going backwards". This
could be even security-relevant since the app running inside the enclave
may behave unexpectedly if time teleports back.

This commit adds a sanity adjustment which guarantees no rewinding time
is applied when the OCALL-obtained timestamp is "back in time" than the
last recorded RDTSC timestamp.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>