[LibOS] Keep main thread alive if it's not the last one
This change is to cover a corner case of a non-main thread performing
`execve()`. Linux does the following: the main (leader) thread is
terminated, and the non-main thread assumes its identity:

    https://elixir.bootlin.com/linux/v6.0/source/fs/exec.c#L1078

Before this commit, Gramine simply terminated the main thread. This led
to the process being marked as a zombie (even though the non-main thread
runs correctly and responds to signals). This in turn confused tools
like `docker kill`, which couldn't find the Gramine process anymore.

This commit fixes this corner case. Gramine can't do as Linux because
there is no way to ask the host OS to "rewire" the identity of one
thread to the other thread. Thus we introduce a workaround of keeping
alive (but in a "parked" state) the main thread -- it sleeps infinitely.

In this case, the main thread and its associated resources (like the
thread's LibOS stack) are not freed -- this leaks memory, but only once
per process, as there is only one main thread per process, even after
several execve invocations.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>