[CI] Do not forward host `PATH` envvar into Docker container
Previously, our CI were overriding the Docker-image `PATH` with the
Jenkins environment (host) `PATH`. If a Docker image is based on one OS
distro (e.g. Ubuntu) and the host has another OS distro (e.g. Debian),
then the `PATH` inside a Docker container could be unexpected. In
particular, Ubuntu has `/sbin` in its PATH but Debian doesn't, and this
led to failures of in-Docker scripts that want to use `/sbin` apps.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>