[LibOS] Use dynamic allocation for `libos_process.cmdline`
Previously Gramine had a static buffer of one page (4KB) in size to hold
cmdline, which turns out to be not enough for Apache Kafka workloads
(which is actually a `/bin/java` command line with many JAR files listed
in the command line).

This commit changes to use dynamic allocation for it and propagates on
fork/clone.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>