squash! [PAL/Linux] Do not call `pal_strerror` before/during/failed PAL relocation
!TODO: use this commit msg:

[PAL] Do not call `pal_strerror` before/during failed PAL relocation

We can't do that because the array which contains error strings
might not be relocated yet. The error strings reside in .rodata
segment. If we use `pal_strerror` and relocations are in use
(default build), the function will return the offset instead of
the real address. This will cause Gramine to segfault.

Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>