[LibOS] Don't allow inlining of `gramine_call()`
`gramine_call()` is a special-syscall function, mainly used to test
Gramine-specific functionality from the app. Previously it was always
inlined in the app. However, an optimized application may end up with
`jmpq *%gs:<gramine-syscall>` instruction in the middle of app-specific
logic. This logic may involve XSAVE registers like XMM/YMM/etc. which
are *not* required to be saved/restored across the `jmpq`. This is in
contrast to the `call` instruction, which forces the compiler to
save/restore XSAVE registers across `call` (according to the Sys-V
x86-64 ABI). So this commit disallows inlining, to produce the `call`
instruction.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>