[LibOS,PAL] Use non-encrypted pipes for `libos_pollable_event`
Gramine has a struct `libos_pollable_event`, which is currently used
only for threads and IPC worker. This object is used purely for
wait/wakeup signaling (so that threads can sleep waiting on such events,
e.g. a thread may wait when this event is triggered during `epoll`) and
does not transfer any data. However, it was previously implemented as
PAL pipes which are transparently encrypted in case of SGX backend. This
implementation is expensive under the following conditions:
* The app creates and destroys many threads during its execution (this
  was noticed on some AI/ML Python workloads).
* The app runs under SGX (in this case, PAL pipes are TLS-encrypted and
  require an expensive TLS handshake at each thread creation).

This commit introduce a new type of PAL pipes called "raw pipes",
together with new URI prefixes to mark the pipe as unencrypted. We then
use such non-encrypted pipes for `libos_pollable_event` under Linux-SGX
PAL.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>