squash! [LibOS,PAL] Use non-encrypted pipes for `libos_pollable_event`
!TODO use this commit msg:

[LibOS,PAL/Linux-SGX] Use non-encrypted pipes for `libos_pollable_event`

Gramine has a struct `libos_pollable_event`, which is currently used
only for threads and IPC worker. This object is used purely for
wait/wakeup signaling (so that threads can sleep waiting on such events,
e.g. a thread may wait when this event is triggered during `epoll`) and
does not transfer any data. However, it was previously implemented as
PAL pipes which are transparently encrypted in case of SGX backend. This
implementation is expensive under the following conditions:
* The app creates and destroys many threads during its execution (this
  was noticed on some AI/ML Python workloads).
* The app runs under SGX (in this case, PAL pipes are TLS-encrypted and
  require an expensive TLS handshake at each thread creation).

This commit extends the support of `PAL_OPTION_PASSTHROUGH` in
`pal_stream_options_t` to pipes on Linux-SGX PAL (Linux PAL simply
doesn't care) -- if this option is NOT set, then the pipes are
encrypted; otherwise they are "passthrough'ed" (i.e., unencrypted).

Signed-off-by: Kailun Qin <kailun.qin@intel.com>