[Pal/Linux-SGX] sgx_gdb: Fix `ptrace(PTRACE_PEEKDATA)` invocation
Previously, Gramine's GDB plugin used `ptrace(PTRACE_PEEKDATA, pid)`
to detect all TIDs of enclave threads. This way of invocation (with a
PID of a process instead of a stopped-tracee thread TID) led to sporadic
errors of ptrace syscall with `-ESRCH`. Ptrace can return this error
code when the specified tracee (the main thread) is not currently
stopped. Because this invocation could happen from child threads, this
error happened when debugging multi-threaded workloads.

The fix is to always use `ptrace(PTRACE_PEEKDATA, tid)` where TID
is the currently ptrace-stopped tracee thread.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>