[tools/sgx] Add gramine-ratls wrapper
This tool generates X.509 certificate and matching private key using
RA-TLS library. It saves those as DER-encoded files under paths given as
first two CLI arguments. If further arguments are passed, those are
interpreted as a command that is then executed using execvp().

It is intended to launch standalone TLS (HTTPS) servers which require
cert and key passed as files.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>