DONTMERGE: [PAL/Linux-SGX] Add encrypted-and-trusted files
Previously, Gramine allowed encrypted files (aka protected files), which
do not have a corresponding hash in the manifest file and thus their
contents are not reflected in MRENCLAVE. Also, Gramine had trusted files
whose contents are reflected in MRENCLAVE but these files are not
encrypted.

This commit allows to combine two functionalities and get
encrypted-and-trusted files as a result. The manifest syntax doesn't
change and looks like this:

  fs.mounts = [
    { type = "encrypted", path = "/enc_file", uri = "file:enc_file" },
  ]
  sgx.trusted_files = [
    "file:enc_file", # or with {uri = ..., sha256 = ...} syntax
  ]

Note that previously Gramine had unintuitive behavior of *ignoring* the
entry in `sgx.trusted_files` if it was already marked as encrypted file.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>