[LibOS] Rename POSIX lock related structs and functions
This is a preparatory commit to introduce flock (BSD locks) support.
Previously, Gramine supported only POSIX locks, which was reflected in
struct and function names (the `posix_` prefix and the `pl`
abbreviation). With BSD locks, the same structs and functions will be
used for both lock types, so this commit renames `posix` to `file` and
`pl` to `file_lock`. Additionally, misleadingly named `struct fs_lock`
is renamed to `struct dent_file_locks` (means: association of a dentry
with its file locks).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>