WIP [CI] Add clone3 to old seccomp in CI and fix pipelines
*DRAFT*

IIUC newer glibc (>= 2.34) attempts to use clone3(2) and faults cleanly
when it returns ENOSYS (like when running on older kernel where there is
no such syscall), but fails badly if the return is EPERM (generated by
seccomp when not permitted).

https://stackoverflow.com/a/72057185

This change makes it so that seccomp returns ENOSYS instead (if caller
does not have CAP_SYS_ADMIN).

The real underlying problem is that we have very old docker and very old
kernel.

Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>