[Meson,CI-Examples] Don't create symlinks in Gramine runtime dir
Gramine runtime dir is used for libc libraries (glibc or musl).
Previously, we also added symlinks to Gramine-specific mbedTLS, RA-TLS
and SecretProv libraries into this runtime dir. However, these latter
libraries are built using the toolchain of the OS distro and thus may
have a different libc than specified in the Gramine runtime dir.
Therefore, the two kinds of libraries should be separated, and no
symlinks should be created.

This is a *breaking* change for applications, because if the app uses
Gramine-specific mbedTLS, RA-TLS or SecretProv libs it is required to
add corresponding entries in the manifest. This commit updates our CI
examples accordingly.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>