[LibOS] Check existence of file during `chroot_stat()`
This commit temporarily opens the file during `stat()` callback to check
if the file still exists on the host. Without this open and if the same
file was opened before, Gramine would serve the `chroot_stat()` callback
from the cached `dent->inode` and wouldn't notice that the file was
removed on the host. Note that such check is relevant only to chroot
files, because all other file types are either pseudo in-memory files
(dev, etc, sys, tmpfs) or fully controlled by Gramine files (encrypted).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>