[common] Do not expand `__FILE_NAME__` to our func call
Previously we had a function that emulated `__FILE_NAME__` macro (which
was introduced only recently in GCC12 and Clang9) by truncating the file
name in a loop. It led to unnecessary calls to this func even when logs
were disabled, ultimately resulting in perf overhead. We observed around
5% overhead because of this on file-I/O intensive benchmarks (because
there we print a lot of log_debug info). This commit simply falls back
to a more clunky `__FILE__` macro when `__FILE_NAME__` is not defined.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>