[PAL] Add `post_callback` function arg to `pal_main()`
This is useful so that the host-specific PAL can perform some
post-initialization actions. In particular, the
`print_warnings_on_insecure_configs()` func in SGX PAL is moved to the
post-initialization phase because this info should be logged into the
`loader.log_file` (if it was specified in the manifest) instead of on
the terminal.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>