squash! [PAL] Fail if Gramine binaries are built with RELR relocs
[PAL] Fail if Gramine binaries are built with unrecognized dyn entries

Gramine has its own dynamic loader to load the SGX PAL and the LibOS
binaries (and also the PAL tests). The dynamic loader is very limited in
its abilities and expects a specially-built binaries. To prevent Gramine
from accidentally allowing important but unaccounted-for dynamic
entries, this commit introduces an allow list of entries.

This commit is a response to RELR (relative relocs) dynamic entries.
RELR is a new type of relocations recently added to compilers/linkers,
e.g. glibc 2.36 and musl 1.2.4. These relocs are not enabled by default;
however, if the system enables them during Gramine build, Gramine would
segfault (before this commit) -- because it ignored RELR relocs, which
led to some symbols being not relocated and ultimately to segfaults.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>