[LibOS,PAL/Linux-SGX] Add EDMM lazy allocation support
The initial implementation of EDMM support pre-accepted every mmapped
region of enclave pages. However in some cases, applications (e.g.,
Java runtime) may rely on lazy allocation of pages, where the VMAs are
reserved but not actually committed to physical memory. In particular,
`mmap(..., MAP_NORESERVE)` requests are used in such cases -- to mmap a
huge chunk of memory (possibly never used in the future) at once and
then commit pages on demand on page fault events. Due to the lack of
lazy allocation concept in Gramine, this incurred performance and
physical memory overhead on allocating and accepting the whole range of
enclave pages.

The commit adds the lazy allocation support by deferring page accepts to
page-fault events when Gramine notices `MAP_NORESERVE` flag in the mmap
request. It also introduces a bit vector indicating the commit status of
each enclave page and an upcall from the SGX EDMM backend into the LibOS
to ask for additional info on a particular page.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>