squash! [PAL/Linux-SGX] Disallow memfault handling if `sgx.require_exinfo = false`
!TODO use below commit msg:

[PAL/Linux-SGX] Disallow memfault handling if no faulting address reported

Previously, Linux-SGX PAL propogated the memory fault exception with
address 0 anyway to LibOS/app even if no faulting address was actually
reported by SGX. This could happen when `sgx.require_exinfo` is disabled
in the manifest or the environment does not support EXINFO at all.
However, LibOS can't/shouldn't handle such case for security reasons.

This commit adds a check on whether the faulting address is reported
before handling a memory fault and we loudly fail if it's not.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>