pal: Implement get_ranges_max() to get the maximum value + 1
Implement get_ranges_max() to get the maxium value + 1 of an entry.
Use this function to get the maximum value + 1 for possible cpus
and nodes. This value will then be used for array sizes.

On one of my systems /sys/devices/system/node/possible shows '1,8'
and therefore we want to get '9' here.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>