pal: Implement get_ranges_max() to get the maximum value + 1
Implement get_ranges_max() to get the maximum value + 1 of an entry.
Use this function to get the maximum value + 1 for possible cpus
and nodes. This value, such as nodes_max, will then be used for
array sizes.

On one of my systems /sys/devices/system/node/possible shows '1,8'
and therefore we want to get '9' here to allocate an array for
9 entries (0 to 8), of which only entries 1 and 8 are used.

Switch the nodes_cnt to be gotten from counting the entries in
/sys/devices/system/node/online so that we count the number of nodes
that are online, which should be '2' for input of '1,8'. This
now ensures that nodes_cnt <= nodes_max and any numa node set online
should then be < nodes_max, thus avoiding out-of-bounds accesses
to the numa node array.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>