[PAL] Implement get_next_online_numa_node() for non-consecutive entries
The current implementation doesn't work if the numa nodes that are online
are non-consecutive, such as when numa online sysfs entry shows '0,8'.
Therefore, implement get_next_online_numa_node() to cover this case and
give us the next online node starting at 'sidx'. 'sidx' is advanced by the
loops so that on first loop '0' is returned and on 2nd loop with sidx
input '1', '8' is returned. We can expect that nodes_cnt of online nodes
among the nodes_max entries will be found, otherwise the sysfs would be
inconsistent.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>