[LibOS,PAL] Refactor LibOS's `/dev/tty` and PAL's corresponding `dev:tty`
In Gramine, similar to UNIX, the stdin, stdout and stderr streams are
file descriptors on the `/dev/tty` device (controlling terminal):
- `stdin = open("/dev/tty", O_RDONLY)`
- `stdout = stderr = open("/dev/tty", O_WRONLY | O_APPEND)`

On the LibOS side, the previous implementation was confusing: it mounted
`/dev/tty` device file in a special way: not via Gramine's devfs, but
instead via chrootfs.

On the PAL side, the `/dev/tty` file was confusingly mapped to PAL's
`dev:tty`, which was implemented via `PAL_TYPE_DEV` but in a special
way: all device APIs in `pal_devices.c` had a special path for
`dev:tty` (which was also erroneously checked by comparing the FD to 0
or 1, even though these host FDs could be different in the child).

This commit refactors this mess:
- In LibOS, `/dev/tty` becomes a pseudo-file mounted via devfs.
- In PAL, `/dev/tty` maps to `console:`, which is a new type of PAL
  handle `PAL_TYPE_CONSOLE`. This new type is removed from
  `pal_devices.c` and instead implemented in a new `pal_console.c`.

New LibOS test `console` is added to test the tty/console.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>