[LibOS,PAL] Rename `PalObjectClose()` to `PalObjectDestroy()`
Previous name didn't make it clear that this PAL API not only closes the
associated host resources, but also destroys/frees the PAL handle object
itself. The handle-type-specific callbacks are also renamed from
`close()` to `destroy()`. Also, this function and all its callbacks are
of type `void` now and never return a value (as it's useless anyway).

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>