[PAL/Linux-SGX] Apply EXINFO to #GP in exception-handling flow
Previously, EXINFO information and `has_hw_fault_address` applied only
to #PF hardware exception. However, this info must be set on #GP
exceptions too. Otherwise Gramine may fail with "Tried to handle a
memory fault with no faulting address ..." message on a #GP (and if the
manifest contains `sgx.use_exinfo = true`).

This bug was found on the `stress-ng-sigsegv` workload: the app
installed a SIGSEGV handler and induced a #GP. Without this commit,
Gramine failed with an above message.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>