[LibOS] Defer put_handle on thread exit for those with async events
It was possible that during thread_exit() a PAL_HANDLE was freed even
though it had an async event associated with it. In some case the
async_worker was still running and had this PAL_HANDLE on its async_list
and ended up using it. Fix this bug by checking whether PAL_HANDLEs in
the call path of thread_exit -> detach_all_fds -> detach_fd have an
async event associated and if this is the case queue them for the
async worker thread to perform the put_handle on and take it off its
async_list. Otherwise put_handle is called immediately (as before).

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>